+++ spotipy/__init__.py
@@ -1,3 +1,4 @@
 from .client import *  # noqa
 from .oauth2 import *  # noqa
 from .util import *  # noqa
+from .exceptions import *  # noqa