+++ spotipy/client.py
@@ -4,8 +4,6 @@
 
 from __future__ import print_function
 
-__all__ = ["Spotify", "SpotifyException"]
-
 import json
 import sys
 import warnings
@@ -14,22 +12,7 @@ import requests
 import urllib3
 import six
 
-
-class SpotifyException(Exception):
-    def __init__(self, http_status, code, msg, headers=None):
-        self.http_status = http_status
-        self.code = code
-        self.msg = msg
-        # `headers` is used to support `Retry-After` in the event of a
-        # 429 status code.
-        if headers is None:
-            headers = {}
-        self.headers = headers
-
-    def __str__(self):
-        return "http status: {0}, code:{1} - {2}".format(
-            self.http_status, self.code, self.msg
-        )
+from spotipy.exceptions import SpotifyException
 
 
 class Spotify(object):